refactored WebScraper namespace to Web
