+++ NVUpdateManager.NotificationService/NVUpdateManager.NotificationService.csproj
@@ -26,7 +26,8 @@
   <ItemGroup>
     <ProjectReference Include="..\NVUpdateManager.Core\NVUpdateManager.Core.csproj" />
     <ProjectReference Include="..\NVUpdateManager.EmailHandler\NVUpdateManager.EmailHandler.csproj" />
-    <ProjectReference Include="..\NVUpdateManager.WebScraper\NVUpdateManager.WebScraper.csproj" />
+    <ProjectReference Include="..\NVUpdateManager.WebScraper\NVUpdateManager.Web.csproj" />
+    <ProjectReference Include="..\NVUpdateManager.Web\NVUpdateManager.Web.csproj" />
   </ItemGroup>
 
   <ItemGroup>
